+++ MonkeyWrench.Web.WebService/MonkeyWrench.Web.WebService.csproj
@@ -90,7 +90,6 @@
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
-    <Compile Include="Notifications.cs" />
     <Compile Include="Upload.cs" />
     <Compile Include="WebServices.asmx.cs">
       <DependentUpon>WebServices.asmx</DependentUpon>
@@ -101,10 +100,14 @@
     </Compile>
     <Compile Include="WrenchBuild.cs" />
     <Compile Include="Maintenance.cs" />
+    <Compile Include="Notifications\Notifications.cs" />
+    <Compile Include="Notifications\IrcNotification.cs" />
+    <Compile Include="Notifications\EmailNotification.cs" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="App_Data\" />
     <Folder Include="Properties\" />
+    <Folder Include="Notifications\" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="web.config" />
